+++ HRLeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,9 @@
 ﻿using AutoMapper;
+using HRLeaveManagement.Application.Exceptions;
+using HRLeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
 using HRLeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Application.Responses;
 using HRLeaveManagement.Domain;
 using MediatR;
 using System;
@@ -24,8 +27,17 @@ namespace HRLeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
 
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
+            var response = new BaseCommandResponse();
             var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+            if (ReferenceEquals(leaveRequest, null))
+            {
+                response.Success = false;
+                response.Message = "LeaveRequest deletion failed.";
+                throw new NotFoundException("LeaveRequest", request.Id);
+            }
+            
             await _leaveRequestRepository.Delete(leaveRequest);
+            response.Message = "LeaveRequest deletion successful.";
             return Unit.Value;
         }
     }
